Added commons text dep, file structure refactored
