+++ src/main/java/ru/ncedu/socialnetwork/api/models/ProjectGhubMod.java
@@ -1,10 +1,10 @@
-package ru.ncedu.socialnetwork.models;
+package ru.ncedu.socialnetwork.api.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Project {
+public class ProjectGhubMod {
     private long id;
     private String name;
     private String description;
